Update ui-automation.yaml - support comment on PR post merge
The original file was written for an on: pull-request trigger. This updates the workflow to actually work with an on: push trigger. The difference being that the trigger is not tied to a PR, so the PR needs to be "found" before the comment can be left.

I would be fine with the pull-request trigger, but the problem with that is that it runs against develop, which isn't helpful at that stage. If we wanted this to run on pull-request, it needs to hit the feature branch.