Update ui-automation.yaml - Use alternative github action to find pull request
This may not work given that we delete the branch after it is merged. It's a matter of timing, I suppose. If the action finds the PR before the branch is deleted, then it may still work. But somehow I doubt it. It's worth a shot before going back and totally rethinking how this is going to work.